+++ boa/src/syntax/ast/node/call.rs
@@ -1,5 +1,3 @@
-//! Expression nodes.
-
 use super::{join_nodes, Node};
 use gc::{Finalize, Trace};
 use std::fmt;
@@ -66,53 +64,3 @@ impl From<Call> for Node {
     }
 }
 
-/// The `new` operator lets developers create an instance of a user-defined object type or of
-/// one of the built-in object types that has a constructor function.
-///
-/// The new keyword does the following things:
-///  - Creates a blank, plain JavaScript object;
-///  - Links (sets the constructor of) this object to another object;
-///  - Passes the newly created object from Step 1 as the this context;
-///  - Returns this if the function doesn't return its own object.
-///
-/// More information:
-///  - [ECMAScript reference][spec]
-///  - [MDN documentation][mdn]
-///
-/// [spec]: https://tc39.es/ecma262/#prod-NewExpression
-/// [mdn]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/new
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Debug, Trace, Finalize, PartialEq)]
-pub struct New {
-    call: Call,
-}
-
-impl New {
-    /// Gets the name of the function call.
-    pub fn expr(&self) -> &Node {
-        &self.call.expr()
-    }
-
-    /// Retrieves the arguments passed to the function.
-    pub fn args(&self) -> &[Node] {
-        &self.call.args()
-    }
-}
-
-impl From<Call> for New {
-    fn from(call: Call) -> Self {
-        Self { call }
-    }
-}
-
-impl fmt::Display for New {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "new {}", self.call)
-    }
-}
-
-impl From<New> for Node {
-    fn from(new: New) -> Self {
-        Self::New(new)
-    }
-}
